payment-service/AOS-Project/Trip_service/requirements.txt
Django==4.2
djangorestframework==3.14
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
Pillow==10.1.0
pika==1.3.2
python-consul==1.1.0
requests==2.31.0